rename iLockScroll -> iLockPageScroll
