+++ src/traits/i-lock-page-scroll/i-lock-page-scroll.ts
